server/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
drf-spectacular>=0.27
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg[binary]>=3.1
celery>=5.3
redis>=4.5
structlog>=24.1.0
sentry-sdk>=1.40.6
pytest>=7.4
pytest-django>=4.7
